Fixed warnings:
Uncaught result for fread() calls.
Type mismatch.

+++ src/cdmf_csr5.cpp
@@ -61,7 +61,7 @@ void cdmf_csr5(SparseMatrix& R, MatData& W_c, MatData& H_c, parameter& param, ch
     std::cout << "[INFO] Initiating OpenCL Time: " << deltaTAB.count() << " s.\n";
 
     for (unsigned t = 0; t < param.k; ++t) {
-        for (unsigned c = 0; c < R.cols; ++c) {
+        for (long c = 0; c < R.cols; ++c) {
             H_c[t][c] = 0;
         }
     }
